requirements.txt
aerich==0.8.1
aiomysql==0.2.0
aioredis==2.0.1
aiosqlite==0.20.0
annotated-types==0.7.0
anyio==4.8.0
async-timeout==5.0.1
asyncclick==8.1.8.0
bcrypt==4.2.1
cffi==1.17.1
click==8.1.8
colorama==0.4.6
crud==0.1
cryptography==44.0.1
dictdiffer==0.9.0
dnspython==2.7.0
ecdsa==0.19.0
email_validator==2.2.0
fastapi==0.115.8
fastapi-jwt-auth==0.5.0
functions==0.7.0
h11==0.14.0
idna==3.10
iniconfig==2.0.0
iso8601==2.1.0
packaging==24.2
passlib==1.7.4
pluggy==1.5.0
pyasn1==0.6.1
pycparser==2.22
pydantic==2.10.6
pydantic-settings==2.7.1
pydantic_core==2.27.2
PyJWT==1.7.1
PyMySQL==1.1.1
pypika-tortoise==0.5.0
pytest==8.3.4
python-dotenv==1.0.1
python-jose==3.3.0
python-multipart==0.0.20
pytz==2025.1
redis==5.2.1
rsa==4.9
schemas==0.7.1
six==1.17.0
sniffio==1.3.1
starlette==0.45.3
tomlkit==0.13.2
tortoise-orm==0.24.0
typing_extensions==4.12.2
uvicorn==0.34.0
